third_party/blink/web_tests/external/wpt/trusted-types/default-policy-callback-arguments-expected.txt

This is a testharness.js-based test.
[FAIL] Test callback arguments, case HTMLScriptElement textContent
  assert_equals: Expecting the sink name. expected "HTMLScriptElement textContent" but got "Node textContent"
Harness: the test ran to completion.
